fix: Resolve CI test failures unrelated to timestamp feature
Fixed compilation and test failures in executor module:

1. Export missing types in iterator module:
   - Added ProjectionIterator and LazyNestedLoopJoin exports
   - Fixes compilation errors in iterator tests

2. Update aggregate test assertions:
   - Changed Float to Numeric in 6 test assertions
   - Aggregate functions now return Numeric instead of Float
   - Fixed floating point precision in window aggregate test

Test results:
- ✅ 626 executor tests passing (was 620 with 6 failures)
- ✅ All compilation errors resolved
- ✅ No test failures related to timestamp parsing changes

These fixes are unrelated to the timestamp parsing feature but were
blocking CI. The timestamp feature tests (27 unit + 19 integration)
continue to pass.

+++ crates/executor/src/tests/aggregate_caching.rs
@@ -92,7 +92,7 @@ fn test_repeated_count_star_cached() {
 
     let result = executor.execute(&stmt).unwrap();
     assert_eq!(result.len(), 1);
-    assert_eq!(result[0].values[0], types::SqlValue::Float(257.0));
+    assert_eq!(result[0].values[0], types::SqlValue::Numeric(257.0));
 }
 
 #[test]
@@ -151,7 +151,7 @@ fn test_repeated_sum_cached() {
     let result = executor.execute(&stmt).unwrap();
     assert_eq!(result.len(), 1);
     // SUM(amount) = 60, so 60 + 60*2 = 60 + 120 = 180
-    assert_eq!(result[0].values[0], types::SqlValue::Float(180.0));
+    assert_eq!(result[0].values[0], types::SqlValue::Numeric(180.0));
 }
 
 #[test]
@@ -268,11 +268,11 @@ fn test_cache_cleared_between_groups() {
 
     // Category A: COUNT(*) = 2, so 2 + 2 = 4
     assert_eq!(result[0].values[0], types::SqlValue::Varchar("A".to_string()));
-    assert_eq!(result[0].values[1], types::SqlValue::Float(4.0));
+    assert_eq!(result[0].values[1], types::SqlValue::Numeric(4.0));
 
     // Category B: COUNT(*) = 3, so 3 + 3 = 6
     assert_eq!(result[1].values[0], types::SqlValue::Varchar("B".to_string()));
-    assert_eq!(result[1].values[1], types::SqlValue::Float(6.0));
+    assert_eq!(result[1].values[1], types::SqlValue::Numeric(6.0));
 }
 
 #[test]
